12670 Change nested try/except in distro_const.py missed during 2.6 python porting 12662 DC error: get_value_from_helper: Helper method deflt_distro_name raised an exception 12521 libaiscf_pymod: needs to be linked against /usr/lib/libscf! 12476 migrate libtransfer to python2.6 12466 Update ManifestServ and finalizer to work with python version 2.6 12464 ICT should be updated to work with Python version 2.6 12424 AI Websever unable to handle manifest with null IPv4 or MAC address criteria if set in at least one manifest 12391 Transition from Python2.4 to Python2.6 - AI client portion 12162 Need to transition the AI webserver to use the SQLite3 module instead of PySQLite2 11680 The libbe Python bridge needs to be compiled against libpython2.6 11679 beadm should be updated to work with Python version 2.6 11607 Should remove "prototype" out of AI commands 8015 list-manifests output is hard to read
--HG--
rename : usr/src/cmd/distro_const/DC_checkpoint.py => usr/src/cmd/distro_const/dc_checkpoint.py
rename : usr/src/cmd/distro_const/DC_defs.py => usr/src/cmd/distro_const/dc_defs.py
rename : usr/src/cmd/distro_const/DC_tm.py => usr/src/cmd/distro_const/dc_tm.py

+++ usr/src/lib/libti_pymod/libti.c
@@ -20,7 +20,7 @@
  */
 
 /*
- * Copyright 2008 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2009 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 #include <Python.h>
@@ -123,7 +123,7 @@ static boolean_t
 add_uint8_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 {
 	uint8_t		*val_array;
-	int		len, index;
+	Py_ssize_t	len, index;
 
 	/*
 	 * Find out how big the list of uint8_ts is.
@@ -174,7 +174,7 @@ static boolean_t
 add_uint16_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 {
 	uint16_t	*val_array;
-	int		len, index;
+	Py_ssize_t	len, index;
 
 	/*
 	 * Find out how big the list of uint16_ts is.
@@ -227,7 +227,7 @@ static boolean_t
 add_uint64_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 {
 	uint64_t	*val_array;
-	int		len, index;
+	Py_ssize_t	len, index;
 
 	/*
 	 * Find out how big the list of uint64_TS IS.
@@ -280,7 +280,7 @@ static boolean_t
 add_boolean_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 {
 	boolean_t	*val_array;
-	int		len, index;
+	Py_ssize_t	len, index;
 
 	/*
 	 * Find out how big the list of booleans is. 
@@ -335,9 +335,9 @@ add_boolean_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 static boolean_t
 add_string_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 {
-	char	**val_array;
-	char	*value;
-	int	len, index;
+	char		**val_array;
+	char		*value;
+	Py_ssize_t	len, index;
 
 
 	/*
@@ -394,7 +394,7 @@ add_nvlist_array(nvlist_t *attrs, char *attribute, PyObject *pvalue)
 	nvlist_t	*fs_attrs;
 	nvlist_t	**val_array;
 	PyObject	*list_obj;
-	int		len, index, i;
+	Py_ssize_t	len, index, i;
 	int		ret = B_TRUE;
 
 	/*
